Fix disk io

+++ dbV1/src/constants.rs
@@ -1,7 +1,7 @@
 use std::mem::size_of;
 pub const COLUMN_USERNAME_SIZE: usize = 32;
 pub const COLUMN_EMAIL_SIZE: usize = 255;
-pub const ID_SIZE: usize = size_of::<usize>() as usize;
+pub const ID_SIZE: usize = size_of::<i32>() as usize;
 pub const USERNAME_SIZE: usize = size_of::<[char; 32]>() as usize;
 pub const EMAIL_SIZE: usize = size_of::<[char; 255]>() as usize;
 pub const ID_OFFSET: usize = 0;
@@ -10,5 +10,5 @@ pub const EMAIL_OFFSET: usize = USERNAME_OFFSET + USERNAME_SIZE;
 pub const ROW_SIZE: usize = ID_SIZE + USERNAME_SIZE + EMAIL_SIZE;
 pub const PAGE_SIZE: usize = 4096;
 pub const TABLE_MAX_PAGES: usize = 100;
-pub const ROWS_PER_PAGE: usize = (PAGE_SIZE / ROW_SIZE) as usize;
+pub const ROWS_PER_PAGE: usize = PAGE_SIZE / ROW_SIZE;
 pub const TABLE_MAX_ROWS: usize = ROWS_PER_PAGE * TABLE_MAX_PAGES;
